fast fixes

+++ migrations/m161119_111911_altertable_faq_category.php
@@ -15,7 +15,7 @@ class m161119_111911_altertable_faq_category extends Migration
     public function safeUp()
     {
 
-        $this->addColumn('{{%faq_category}}','sort',$this->integer(1)->null());
+        $this->addColumn('{{%faq_category}}','sort',$this->integer(10)->null());
     }
 
     public function safeDown()
